+++ vial-plugin/http/tests.py
@@ -1,7 +1,7 @@
 from textwrap import dedent
 
 from .util import (parse_request_line, render_template, get_headers_and_templates,
-                   find_request)
+                   find_request, pretty_xml)
 
 
 def fill(param):
@@ -108,3 +108,18 @@ def test_find_request():
     assert find_request(lines, 11) == ('POST /uri5', 'boo', 13)
     assert find_request(lines, 12) == ('POST /uri5', 'boo', 13)
     assert find_request(lines, 13) == ('POST /uri5', 'boo', 13)
+
+
+def test_pretty_xml():
+    from cStringIO import StringIO
+    buf = StringIO()
+
+    pretty_xml('''\
+        <root xmlns:d='http://boo' xmlns="http://boom">
+            <d:child d:foo="boo">some text&gt;<child2 xmlns:foo='http//foo'>text</child2>another text</d:child>
+            <d:child>boo</d:child>
+            <d:child>foo</d:child>
+            <d:child>zoo</d:child>
+            <d:child>hoo</d:child>
+        </root>''', buf)
+    print buf.getvalue()